requirements.txt
sentence-transformers
scikit-learn==1.3.2
Pillow
requests
boto3
streamlit
pandas
torch